requirements.txt
tensorflow-cpu
opencv-python-headless
streamlit
numpy
matplotlib
scikit-learn
